Implemented heuristicsDialogSlice
