--- src/constants/dialog/heuristicsDialogActionTypes.js
@@ -1,6 +1,0 @@
-const heuristicsDialogActionTypes = {
-	CLOSE: 'CLOSE_HEURISTICS_DIALOG',
-	OPEN: 'OPEN_HEURISTICS_DIALOG'
-};
-
-export default heuristicsDialogActionTypes;
